+++ CHANGELOG.md
@@ -8,7 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added links to comment-based help.
-- Refreshed updatable help package.
+- Added `-AttachmentFolderPath` parameter to Build-ReportArchive
+- Added parameters and variables to functions that changed due to new parameter
 
 ## [0.4.0] - 2023-04-07